ShipX v1.30.3
Added global styling and best practices to apply the DRY principle and
improve code readability and reusability. Added an icons folder to
segregate svg icons. As well as the styles folder where it contains the
styling for certain tags and the theme guide as well. Started applying
those for the desktop version of the footer for now.