ShipX v2.0
New components for the entire page redesign will be under
components/library. Also a component showcase page will be added and
only people who know the path can access it. Header design is updated
based on the new update, desktop version is being started. As well as
adding a nav color to the global styles.