ShipX v2.1.0
Added an app button that adjusts if it's a link or a regular/submit
button. Added 2 app demos in the component showcase. As well as the
updated login modal and the updated logo component. While also updated
the Header component and the button styles component to make room for
the second button design based on the learn more buttons in figma.