ShipX v2.5.0
Updated app button. Component showcase updated as well. new components added for the services section and the our numbers section and made them responsive as well. The homepage hero section is also updated, as well as adding the numbers section and the services section that are also responsive. And the header hover functionality as well. And the updated themeGuide and button and badge style variants