ShipX v2.9.0
Added a new dropdown component and updated the tab switcher component. Added a new toolkit section, updated the hero, numbers and services section. As well as removing jsx files that have not been used anymore for version 2. Updated the header functionality. Made some minor adjustments to the styling for badges and the inputs.