ShipX v2.12.0
Updated the app button functionality, as well as the login modal and logistics card components. Added a new office modal component for when app button is clicked it shows the complete location. And moved the partner grid from partners section to a component itself. Updated the hero section background image function. Added a new offices section and as well as updated the partners, platform and services sections. Also added a few new constant variables in the constants.js file to be used in the sections for the homepage.